+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   - `EchoRawSlice()` now supports tracing
   - `EchoSlice()` now supports tracing
   - `Grep()` now supports tracing
+  - `GrepV()` now supports tracing
   - `TestEmpty()` now supports tracing
   - `TestFilepathExists()` now supports tracing
   - `TestNotEmpty()` now supports tracing